The journey was long, but I think we're almost there.
It's my first commit, when I write more than 10 words :rofl:

About a year ago, I came up with the idea of creating an open-source online diary for students. Sure, it was a better time, than now, but I think current events in the world will end anyway. Initially, openSchool was intended only as a diary app for my school in Moscow (my native city) since the educational system there is extremely buggy and does not work well. In spite of this, I thought it would be awesome to expand openSchool to the worldwide community, and so I am writing this commit now.

It wasn't easy to develop this software from scratch, but I hope that the first version will be released by 1st September 2022. There were only two experienced developers: me and Timofey Kniga. To my biggest regret, Timofey decided to left the project, I'm not sure why, but he left the project. Other members of our team work on things like small bug fixes or activities other than coding. Accordingly, I am the only developer. Due to the number of people left in this project, it is sometimes hard to do much. Yes, my activity in opensch/core isn't very significant, but we have a closed repository with a mobile app that takes a lot of time to develop.

Why is the mobile app repository closed? It's simple: when we decided to do openSchool, the plan wasn't to make it open to the entire world. As a result, the source code of the app contains tokens and private keys that we don't want to share. With the first release or earlier, we'll open the source code of our app.

Sincerely,
openSchool team.